% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordenar_vectores.R
\name{ordenar_vectores}
\alias{ordenar_vectores}
\title{Distancias ordenadas entre arboles}
\usage{
ordenar_vectores(id, sp, size, x1, y1)
}
\arguments{
\item{id}{vector del identificador numerico del arbol}

\item{sp}{vector de las especies de los arboles}

\item{size}{vector con las dimensiones del diametro de los arboles}

\item{x1}{vector de las coordenadas x de los arboles}

\item{y1}{vector de las coordenadas y de los arboles}
}
\value{
Regresa cuatro objetos en forma matricial de las distancias, especies, id's y el tamano de los arboles
}
\description{
La función calcula la distancias euclidianas entre arboles y las ordenas en orden ascendente con el metodo de inserccion
}
