% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamano_indice_arbol.R
\name{tamano_indice_arbol}
\alias{tamano_indice_arbol}
\title{Indice de diferenciacion de tamanos}
\usage{
tamano_indice_arbol(id, sp, size, x1, y1, k)
}
\arguments{
\item{id}{vector del identificador numerico del arbol}

\item{sp}{vector de las especies de los arboles}

\item{size}{vector con las dimensiones del diametro de los arboles}

\item{x1}{vector de las coordenadas x de los arboles}

\item{y1}{vector de las coordenadas y de los arboles}

\item{k}{valor numerico que identifica el numero de vecinos cercanos}
}
\value{
Regresa un data.frame con los id's, especies, el tamano y el indice de diferenciacion de tamano de los arboles
}
\description{
La función calcula el índice de diferenciacion de tamano para los arboles que se encuentran en el sitio
}
